refactor: abstract authentication process
Add an interface to a "Composable repository" with login buttons
